requirements.txt
presidio-analyzer==2.2.355
presidio-anonymizer==2.2.355
spacy>=3.7.0
transformers>=4.43.0
torch>=2.2.0
regex
pyyaml
tabulate